Update Dockerfile
Switched base image inside the dockerfile. Up to now, a alpine based
image was chosen to make the docker build as small as possible. However,
the oldest available alpine based image (that is still vulnerable to
cve-2016-3427) runs on jre7. All alpine build running on jre8 are no
longer vulnerable. Usage of jre7 can be annoying, as it is per default
no longer supported by most distors and the tonka-bean will not work
when compiled for jre8.

Therefore, a non-alpine based image was now chosen. This is vulnerable
to cve-2016-3427 and runs jre8. However, unfortunately it takes more
disk space.