Remove non required dependency
Removed the not required dependency to com.sun.net.httpserver. This one
is not part of the Java public API but is present in almost all Java
runtimes including openjdk.

In general the dependency seems not to be an optimal solution. If
someone reports errors after this change, we might include another
dependency for the HTTP stuff.