Improve argument handling during invoke action
This commit aims to resolve #11 by refactoring the argument handling
during the invoke operation. The expected arguments for the operation
were changed and the invoke action now expects a dedicated function
signature to be specified. This allows beanshooter to obtain the type
names from the function signature and to marshall the user specified
arguments correctly during a call. Furthermore, an autowrapping
mechanism was implemented that wraps primitive types into their
corresponding wrapper classes automatically if required. This allows
users now to specify e.g. an int argument just as a plain number.