Add additional operations
This commit adds several additional operations. The info and attr
actions were added within the general operations and allow to obtain
information on the available attributes and methods of an MBean.
Furthermore, attributes can also be modified and searched for specific
keywords by using this action.

The attr action is basically a replacement for the invoke action when
targeting getter methods. The info action is also available on the MBean
level.

Moreover, three new default MBeans were implemented (HotSpot,
DiagnosticCommand and FlightRecorder). These provide some useful
offensitive capabilities and the corresponding functions are now
implemented by default.

All new features are still in a raw state and need to be tested.