Remove newly created role during tomcat write
When using the tomcat write action, beanshooter did not remove the newly
created role. Despite resetting the UserDatabase to the old location,
the role was then still visible as an available MBean.

Furthermore, the payload for the tomcat write action is now saved within
the role description. This makes it easier to cleanup the role, as we
can choose a more simple string as role name.