Simplify poms
During the preparation of the release candidates, the pom files were
adpoted to show the correct version number. However, these changes were
never reflected in develop and the pom related commits on master were
treated as the current and newer version. Therefore, they still pointed
at the release candidate version and we needed to update the pom files.
To add at least some work in this commit, we simplified the poms by
removing the unused namespaces.