Update docker containers
Updated the docker containers. Apart from updating them to the latest
Java versions, the hostname of the containers was changed from
iinsecure.dev to iinsecure.example. According to RFC 2606 this top level
domain name cannot be registered and accidental routing of traffic to
the internet should be prevented. Related to #23