Catch CertPathValidatorException
CertPathValidatorException is raised when the servers TLS settings
violate the current java.security policy. Circumventing such a situation
at runtime is rather difficult. Therefore, we just display an error
message that explains the issue. To solve the issue, the user can edit
the java.security policy and allow the TLS settings configured for the
targeted server. Addresses #24