Make beanshooter Java16+ compatible
beanshooter accesses internal Java modules which was fine for older Java
versions like 8 or 11, but causes exceptions on newer ones like 16 or
17. These errors can be prevented by adding additional command line
arguments to open access to the corresponding modules. As it turns out,
opening modules can also be done within the Jar manifest, which is now
implemented in beanshooter.