Add additional exception handling for standard
When using the standard action, a compiled class file is send to the
application server as part of a TemplatesImpl payload. This class file
is dynamically compiled using the locally used version of Java. If this
version is too new, the application server may refuse to load the
precompiled class. This exeception is now caught and an error msg should
explain the issue.